feat: schedule recurring jobs in prod via symfony/scheduler
app:tasks:reminders:dispatch and app:notifications:dispatch-digest had
documented cadences but nothing ran them in production. The schedule now
lives in-process: App\Scheduler\MainScheduleProvider (#[AsSchedule])
fires DispatchTaskReminders every 5 min and DispatchNotificationDigest
hourly at :55 / daily at 08:00 UTC through a scheduler_default transport
that the existing Messenger worker consumes alongside async — no system
cron, no extra container.

- Command bodies extracted into TaskReminderDispatcher /
  NotificationDigestDispatcher services shared by the thin console
  commands (kept for manual runs) and the new message handlers.
- Schedule is stateful (cache.app) + locked + processOnlyLastMissedRun,
  so the worker's hourly --time-limit recycle can't skip a tick and
  downtime collapses to one catch-up run.
- Worker command updated in compose.yaml and the Helm worker
  Deployment (compose.prod.yaml inherits it).
- New deps: symfony/scheduler + dragonmantank/cron-expression.
- when@test untouched: scheduler_default is materialised only when a
  worker consumes it, which never happens in tests.
- Docs: deployment.md scheduled-jobs section rewritten, job-queue.md
  gains a recurring-jobs guide, CLAUDE.md entry updated.
