+++ api/composer.json
@@ -10,6 +10,7 @@
         "doctrine/doctrine-bundle": "^3.2",
         "doctrine/doctrine-migrations-bundle": "^4.0",
         "doctrine/orm": "^3.0",
+        "dragonmantank/cron-expression": "^3.4",
         "intervention/image": "^4.0",
         "league/flysystem-bundle": "^3.3",
         "minishlink/web-push": "^10.0",
@@ -37,6 +38,7 @@
         "symfony/property-info": "7.4.*",
         "symfony/rate-limiter": "7.4.*",
         "symfony/runtime": "7.4.*",
+        "symfony/scheduler": "7.4.*",
         "symfony/security-bundle": "7.4.*",
         "symfony/serializer": "7.4.*",
         "symfony/twig-bundle": "7.4.*",
