+++ docs/developer/job-queue.md
@@ -20,6 +20,7 @@ email was sent inline, putting SMTP latency on the request's critical path.
 | Queue table | `messenger_messages` (created by `Version20260606130000`) |
 | Async email message | `Symfony\Component\Mailer\Messenger\SendEmailMessage` (framework-provided) |
 | Email producers | every `*Mailer` service / controller that calls `MailerInterface::send()` (e.g. [NotificationDispatcher.php](../../api/src/Service/NotificationDispatcher.php), `InviteMailer`, `PasswordController`) |
+| Recurring-job schedule | [api/src/Scheduler/MainScheduleProvider.php](../../api/src/Scheduler/MainScheduleProvider.php) (symfony/scheduler, consumed as the `scheduler_default` transport) |
 | Worker (dev) | `worker` service in [compose.yaml](../../compose.yaml) |
 | Worker (k8s) | [helm/api-platform/templates/worker-deployment.yaml](../../helm/api-platform/templates/worker-deployment.yaml) |
 
@@ -41,7 +42,11 @@ email was sent inline, putting SMTP latency on the request's critical path.
 - **Tests run synchronously** — under `when@test` the `async` transport is
   `sync://`, so dispatched jobs are handled in-process during the same
   request. Existing email assertions (`NotificationTest`,
-  `NotificationTriggersTest`, …) keep working without a worker.
+  `NotificationTriggersTest`, …) keep working without a worker. The
+  scheduler transport needs no test override: `scheduler_default` isn't a
+  configured transport at all — it's materialised on demand by the
+  scheduler component when a worker consumes it, which never happens in
+  the test environment.
 
 ## Running the worker
 
@@ -56,11 +61,12 @@ docker compose logs -f worker
 Or run a consumer by hand inside the php container:
 
 ```bash
-docker compose exec php bin/console messenger:consume async -vv
+docker compose exec php bin/console messenger:consume async scheduler_default -vv
 ```
 
 **Production (Helm):** a dedicated `*-worker` Deployment runs
-`messenger:consume`. Toggle/scale it via `values.yaml`:
+`messenger:consume async scheduler_default`. Toggle/scale it via
+`values.yaml`:
 
 ```yaml
 worker:
@@ -72,6 +78,60 @@ worker:
 
 The worker handles `SIGTERM` gracefully (finishes the in-flight message,
 then exits), so rolling restarts don't drop jobs.
+
+## Recurring jobs (symfony/scheduler)
+
+Cron-style jobs ride the same worker instead of a system cron:
+[MainScheduleProvider](../../api/src/Scheduler/MainScheduleProvider.php)
+declares the schedule with `#[AsSchedule]`, which exposes it as a virtual
+Messenger transport named `scheduler_default`. When the worker consumes
+that transport, the scheduler computes which triggers are due and yields
+their messages into the normal pipeline — handlers, retries, and the
+failed transport all behave exactly like any other job.
+
+Current schedule (all times UTC):
+
+| Message | Trigger | Handler delegates to |
+| --- | --- | --- |
+| `App\Message\DispatchTaskReminders` | `*/5 * * * *` | `App\Service\TaskReminderDispatcher` |
+| `App\Message\DispatchNotificationDigest('hourly')` | `55 * * * *` | `App\Service\NotificationDigestDispatcher` |
+| `App\Message\DispatchNotificationDigest('daily')` | `0 8 * * *` | `App\Service\NotificationDigestDispatcher` |
+
+The same services back the `app:tasks:reminders:dispatch` and
+`app:notifications:dispatch-digest --period=hourly|daily` console
+commands, so a one-off manual run goes through identical code. Both
+passes are idempotent — reminders dedupe on a (recipient, task, offset)
+unique index, digests stamp `digestedAt` — so an extra run is harmless.
+
+Semantics declared on the schedule:
+
+- **`stateful(cache.app)`** — last-run timestamps survive worker restarts
+  (the worker recycles hourly via `--time-limit`), so a tick that lands
+  during the restart window is caught up instead of skipped.
+- **`processOnlyLastMissedRun(true)`** — after longer downtime, only the
+  most recent missed tick per message is replayed, not the whole backlog.
+- **`lock()`** — two consumers of `scheduler_default` won't process the
+  same tick twice. With the default `LOCK_DSN=flock` and filesystem cache
+  this only holds per host: scale the worker by keeping one replica on
+  `scheduler_default` (others on `async` only) or move `LOCK_DSN` and
+  `cache.app` to shared stores.
+
+Inspect what's due when with:
+
+```bash
+docker compose exec php bin/console debug:scheduler
+```
+
+**Adding a recurring job:** write a message + handler exactly as in
+[Adding a new job](#adding-a-new-job) (skip the routing — the scheduler
+transport delivers it directly), then attach a `RecurringMessage` to the
+schedule in `MainScheduleProvider::getSchedule()`. Keep the handler
+idempotent: under at-least-once delivery plus catch-up runs, "ran twice"
+must be safe.
+
+**Tests:** the schedule never fires in the test environment (nothing
+consumes `scheduler_default` there), and the underlying services are unit-
+testable directly — no override in `when@test` is needed.
 
 ## Inspecting failures
